Add files via upload
![NeomView](https://github.com/aredoteam/NeomView/blob/main/neom-vew-by-aredo.io.png)
Android Studio application to convert the website into an application "NeomView" is a comprehensive application that converts your website into an Android Studio application. The application includes innovative and customizable features, such as:

Three customizable pages: These pages appear when the application is opened for the first time. You can modify and customize the content of these pages according to your needs.

Customizable startup page: This page appears each time the application is opened. You can customize it based on your preferences and project requirements.

Distinct and customizable bottom menu: The application features a visually appealing and customizable bottom menu. You can customize this menu and add links or functions as desired.

Fingerprint support: The application allows for locking access with the user's fingerprint. This feature ensures security and privacy for the user.

App lock with a password: Users can set a password to lock the application, enhancing security and preventing unauthorized access.

Built-in barcode reader: The application includes a built-in barcode reader that allows users to open links within the app while verifying the connection's SSL authenticity.

Customizable no internet connection page: The application includes a customized page that appears when there is no internet connection available. You can customize this page to provide relevant information or instructions.

Please note that the translation is done to the best of my ability, but it may require further editing or adaptation based on specific context or preferences.

Installation:

Import the application into Android Studio. Import the required libraries for the application. Update the page links to match your website. Go to the manifest file and specify the necessary permissions, removing any unnecessary ones. To change the icons, you can replace the icons in the drawable folder. To change the application name, modify the string file. Please note that these instructions are provided in a general context. The specific steps may vary depending on the structure and requirements of your application.

Creative Commons Attribution-NonCommercial [2.0.1] International License (CC BY-NC [2.0.1]) This application is licensed for personal, non-commercial use only. You are not allowed to use this application for commercial purposes without obtaining prior written permission from [AredoTeam]